requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.3
django-cors-headers==3.7.0
django-import-export==2.5.0
djangorestframework==3.12.4
et-xmlfile==1.1.0
idna==2.10
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
python-decouple==3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
sqlparse==0.4.1
tablib==3.0.0
upgrade-requirements==1.7.0
urllib3==1.26.5
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
